Updates viewnames in function tests
