+++ keystone_api/tests/function_tests/stats/test_grant_detail.py
@@ -1,15 +1,15 @@
-"""Function tests for the `stats:request-list` endpoint."""
+"""Function tests for the `stats:grant-list` endpoint."""
 
 from django.urls import reverse
 from rest_framework import status
 from rest_framework.test import APITestCase
 
-from apps.allocations.factories import AllocationRequestFactory
+from apps.research_products.factories import GrantFactory
 from apps.users.factories import MembershipFactory, UserFactory
 from apps.users.models import Membership
 from tests.function_tests.utils import CustomAsserts
 
-VIEW_NAME = 'stats:request-list'
+VIEW_NAME = 'stats:grant-detail'
 
 
 class EndpointPermissions(CustomAsserts, APITestCase):
@@ -54,8 +54,8 @@ class EndpointPermissions(CustomAsserts, APITestCase):
         )
 
 
-class TeamRecordFiltering(APITestCase):
-    """Test returned metrics are filtered by user team membership."""
+class TeamGrantFiltering(APITestCase):
+    """Test returned grant metrics are filtered by user team membership."""
 
     endpoint = reverse(VIEW_NAME)
 
@@ -66,33 +66,33 @@ class TeamRecordFiltering(APITestCase):
         self.team_1 = membership_1.team
         self.team_1_user = membership_1.user
         self.team_1_records = [
-            AllocationRequestFactory(team=self.team_1) for _ in range(2)
+            GrantFactory(team=self.team_1) for _ in range(2)
         ]
 
         membership_2 = MembershipFactory(role=Membership.Role.MEMBER)
         self.user_2 = membership_2.user
         self.team_2 = membership_2.team
         self.team_2_records = [
-            AllocationRequestFactory(team=self.team_2) for _ in range(3)
+            GrantFactory(team=self.team_2) for _ in range(3)
         ]
 
         self.staff_user = UserFactory(is_staff=True)
         self.all_records = self.team_1_records + self.team_2_records
 
     def test_generic_user_statistics(self) -> None:
-        """Verify general users are only returned statistics from teams they are a member of."""
+        """Verify general users only receive statistics for their teams."""
 
         self.client.force_authenticate(self.team_1_user)
         response = self.client.get(self.endpoint)
 
         stats = response.json()
-        self.assertEqual(len(self.team_1_records), stats["request_count"])
+        self.assertEqual(len(self.team_1_records), stats["grant_count"])
 
     def test_staff_user_statistics(self) -> None:
-        """Verify staff users are returned aggregated statistics across all teams."""
+        """Verify staff users receive statistics aggregated across all teams."""
 
         self.client.force_authenticate(self.staff_user)
         response = self.client.get(self.endpoint)
 
         stats = response.json()
-        self.assertEqual(len(self.all_records), stats["request_count"])
+        self.assertEqual(len(self.all_records), stats["grant_count"])
